+++ lab1/src/Drawings/CoordinateSpace/CoordinateSpace.java
@@ -3,10 +3,10 @@ package Drawings.CoordinateSpace;
 import Drawings.graphicElements.Drawable;
 
 public interface CoordinateSpace {
-    public int[][] convert(int[][] userCoordinates);
-    public int convertLengthX(int num);
-    public int convertLengthY(int num);
-    public int[] getUserData();
+    public int[][] convert(float[][] userCoordinates);
+    public int convertLengthX(float num);
+    public int convertLengthY(float num);
+    public float[] getUserData();
     public void setReal(int realMaxX, int realMaxY);
     public Drawable getGrid();
     public void countMouseCoordinates(Mouse mouse);
